Fix parameters so things actually work in Dashing.
This change ensures that we declare parameters before trying
to use them; otherwise, "get_parameters" will fail on startup.
To use this new system, we also slightly change how we deal
with axes that are unassigned; instead of having no parameter,
we now declare the parameter with the mapping set to -1 (an
invalid number).  Finally, we fix up a compiler warning when
printing an int64.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>